backend/requirements.txt
# Lider Garant Django Backend

Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
python-dotenv==1.0.1
Pillow==10.3.0
